dwarf: Fix ARM/THUMB symbol lookup
To find a DWARF debug info, it makes sure by finding symbol using the
address found from the DWARF.  But it failed on ARM (functions that
built for THUMB mode) since it has one-off symbol address (according to
the spec).

To work around it add a 1 to the offset when looking up symbols.  It
should have no problem on other archs because it should be still in the
same symbol (unless it's 0-sized).  After finding a symbol, use address
of the symbol instead since it's the one to be used for lookup.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>