trigger: Pass LP64 info when parsing argspec
It should support to read data from a machine has different endian/
class.  It means that it's possible on x86 reading data recorded
on an ARM board.

However as argspec didn't write full size info for default types,
there's a mismatch if the size is different between record and replay.
Fix it by passing LP64 info to arguments/trigger setup functions to
determine the size of the field properly.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>