mcount: Get rid of mcount_active count
It's a not-so-small performance penalty to change a global variable in
the fast path.  As it's only to prevent race for the finish trigger,
let's use a different approach.

In fact, a signal handler should not do something dangerous such as
acquiring a lock.  So add mcount_trace_finish() to do nothing but
setting the FINISH flag.  Next time any thread gets into the libmcount
it'll see the flag during mcount_guard/unguard_recursion() and will call
mtd_dtor().

The mcount_finish_once() is called when it first saw the flag set or
mcount_finish() will call it at the end.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>